+++ src/main/java/org/truffleruby/language/dispatch/CachedBoxedDispatchNode.java
@@ -108,7 +108,7 @@ public class CachedBoxedDispatchNode extends CachedDispatchNode {
     public String toString() {
         return StringUtils.format(
                 "CachedBoxedDispatchNode(:%s, %x, %s)",
-                getCachedNameAsSymbol().toString(),
+                getCachedName().toString(),
                 expectedShape.hashCode(),
                 method == null ? "null" : method.toString());
     }