+++ CHANGELOG.md
@@ -76,6 +76,7 @@ Compatibility:
 * Implemented `Enumerator::Chain`, `Enumerator#+`, and `Enumerable#chain` (#1859, #1858).
 * Implemented `Thread#backtrace_locations` and `Exception#backtrace_locations` (#1556).
 * Implemented `rb_module_new`, `rb_define_class_id`, `rb_define_module_id`, (#1876, @chrisseaton).
+* Cache the `Symbol` of method names in call nodes only when needed (#1872).
 
 Performance: